+++ CHANGELOG.md
@@ -1,13 +1,17 @@
 # `SwipProteomics/CHANGELOG.md`
 
+#### Mon Sep 28 10:42:56 EDT 2020
+* PSM report was not correct. Somehow it was missing reporter ion abundances. 
+  Data was re-expored as an excel worksheet from PD by GW.
+* Working on re-analysis of the data with MSstatsTMT.
+
 #### Thu Sep 17 10:56:29 EDT 2020
-* install NBGOF package - need to evaluate GOF of NB model
-* added PSM report to data
+* Install NBGOF package - need to evaluate GOF of NB model.
+* Added PSM level data from DPMSR as `data/5539_PSM_Report.txt.gz`.
 
 #### Tue Sep 15 09:33:40 EDT 2020
 * generate sample pca showing good resolution of different subcellular fractions
 * mkdir `docs/review` to init work on reviews
 * init change log file
-